requirements.txt
feedparser==6.0.11
anthropic==0.42.0
pyyaml==6.0.2
requests==2.32.3
